+++ src/components/ProviderInitializer/helpers/index.ts
@@ -6,3 +6,4 @@ export * from './processModifiedAccount';
 export * from './getModifiedLoginToken';
 export * from './getMetamaskProvider';
 export * from './getPasskeyProvider';
+export * from './handleGuardianWarning';