+++ src/utils/storage/local.ts
@@ -2,7 +2,9 @@ import { getUnixTimestamp } from 'utils/dateTime';
 
 export const localStorageKeys = {
   loginExpiresAt: 'sdk-dapp-login-expires-at',
-  logoutEvent: 'sdk-dapp-logout-event'
+  logoutEvent: 'sdk-dapp-logout-event',
+  guardianBreachToastDismissTimestamp:
+    'sdk-dapp-guardian-breach-toast-dismiss-timestamp'
 } as const;
 
 type LocalValueType = keyof typeof localStorageKeys;